Example of using KeyEncryptionAlgorithm OID: 1.2.840.113549.1.1.7 (RSAES_OAEP)
Health.Direct.Common must be .NET Core 5 or greater to control the RSAEncryptionPadding algorithm.  See the `SMIMECryptographer` in the CreateEncryptedEnvelope method for an example of the before and after code.  See the Encrypt_Then_Decrypt_LogsDebugMessages Test for an example of encrypting and decrypting and some logging evidence of the padding used.